pyproject.toml
[project]
name = "home-master"
version = "0.1.0"
description = "муж на час"
authors = [
    {name = "Grigory Kalinin",email = "grigorykalinin@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pylint-django (>=2.6.1,<3.0.0)",
    "django-phonenumber-field (>=8.1.0,<9.0.0)",
    "phonenumbers (>=9.0.10,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pylint]
load-plugins = "pylint_django"